Fix some state issues
