+++ server/src/index.js
@@ -130,10 +130,6 @@ i18next.use(Backend).init({
 
 app.use(rootRouter, requestRateLimiter)
 
-app.all('*', (req, res) => {
-  res.redirect('/')
-})
-
 // eslint-disable-next-line no-unused-vars
 app.use((err, req, res, next) => {
   console.error('Express Error:', err.message)
