+++ src/components/App.jsx
@@ -8,8 +8,8 @@ import {
   createHttpLink,
 } from '@apollo/client'
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
+import ApolloLinkTimeout from 'apollo-link-timeout'
 
-import AbortableLink from '@classes/AbortableLink'
 import UIcons from '@services/Icons'
 import Fetch from '@services/Fetch'
 import Auth from './layout/auth/Auth'
@@ -17,9 +17,11 @@ import Login from './layout/auth/Login'
 import RouteChangeTracker from './RouteChangeTracker'
 import Errors from './Errors'
 
+const timeoutLink = new ApolloLinkTimeout(10000) // 10 second timeout
+
 const client = new ApolloClient({
   uri: '/graphql',
-  link: new AbortableLink().concat(createHttpLink()),
+  link: timeoutLink.concat(createHttpLink({ uri: '/graphql' })),
   cache: new InMemoryCache({
     typePolicies: {
       Query: {
